PWX-3927: Fix unconditional PX-OCI restarts

+++ px-oci-mon/main.go
@@ -1,10 +1,15 @@
 package main
 
 import (
+	"bytes"
 	"fmt"
+	"io"
 	"os"
 	"os/exec"
+	"regexp"
+	"strconv"
 	"strings"
+	"sync"
 	"syscall"
 
 	"github.com/Sirupsen/logrus"
@@ -19,6 +24,13 @@ const (
 	baseDir            = "/opt/pwx/oci"
 	baseServiceName    = "portworx"
 	baseServiceFileFmt = "/etc/systemd/system/%s.service"
+)
+
+var (
+	// rsyncRegex extracts '# files transferred' from rsync's stat/stat2 report
+	rsyncRegex = regexp.MustCompilePOSIX(`Number of regular files transferred: ([0-9,]+)`)
+	// xtractKubeletRegex extracts /var/kubelet -override from running kubelet daemon
+	xtractKubeletRegex = regexp.MustCompile(`\s+--root-dir=(\S+)`)
 )
 
 // usage borrowed from ../../porx/cmd/px-runc/px-runc.go -- TODO: Consider refactoring !!
@@ -61,7 +73,7 @@ examples:
 	os.Exit(1)
 }
 
-func runExternal(name string, params ...string) error {
+func runExternalWithOutput(out io.Writer, name string, params ...string) error {
 	args := make([]string, 0, 4+len(params))
 	args = append(args, "/usr/bin/nsenter", "--mount="+mntFileName, "--", name)
 	args = append(args, params...)
@@ -69,11 +81,54 @@ func runExternal(name string, params ...string) error {
 	logrus.Info("> run: ", strings.Join(args[3:], " "))
 	logrus.Debugf(">>> %+v", args)
 	cmd := exec.Command(args[0], args[1:]...)
-	cmd.Stdout, cmd.Stderr = os.Stdout, os.Stderr
+	if out == nil {
+		cmd.Stdout, cmd.Stderr = os.Stdout, os.Stderr
+	} else {
+		// note: exec.CombinedOutput() assigns to bytes.buffer, like we do
+		cmd.Stdout, cmd.Stderr = out, out
+	}
 	return cmd.Run()
 }
 
-func installPxFromOciImage(di *utils.DockerInstaller, imageName string, cfg *utils.SimpleContainerConfig) error {
+func runExternal(name string, params ...string) error {
+	return runExternalWithOutput(nil, name, params...)
+}
+
+// Output filters --
+
+type regexFilter struct {
+	re  *regexp.Regexp
+	val []byte
+	sync.Mutex
+}
+
+func (r *regexFilter) Write(b []byte) (int, error) {
+	if m := r.re.FindAllSubmatch(b, -1); len(m) > 0 && len(m[0]) > 1 {
+		ll := len(m[0][1])
+		r.Lock()
+		r.val = make([]byte, ll)
+		copy(r.val, m[0][1])
+		r.Unlock()
+	}
+	return os.Stderr.Write(b)
+}
+
+type cachingOutput struct {
+	bb bytes.Buffer
+}
+
+func (c *cachingOutput) Write(b []byte) (int, error) {
+	c.bb.Write(b)
+	return os.Stderr.Write(b)
+}
+
+func (c *cachingOutput) String() string {
+	return c.bb.String()
+}
+
+// -- Output Filters
+
+func installPxFromOciImage(di *utils.DockerInstaller, imageName string, cfg *utils.SimpleContainerConfig) (bool, error) {
 	logrus.Info("Downloading Portworx...")
 
 	if err := di.PullImage(imageName); err != nil {
@@ -87,16 +142,30 @@ func installPxFromOciImage(di *utils.DockerInstaller, imageName string, cfg *uti
 		logrus.WithError(err).Warn("Unable to create pwx directories directly -- retry via shell")
 		err = runExternal("/bin/sh", "-c", "mkdir -p /opt/pwx /etc/pwx")
 		if err != nil {
-			return fmt.Errorf("Unable to create pwx directories: %s", err)
+			return true, fmt.Errorf("Unable to create pwx directories: %s", err)
 		}
 	}
 
+	reFilter := &regexFilter{re: rsyncRegex}
 	err := di.RunOnce(imageName, ociInstallerName, []string{"/opt/pwx:/opt/pwx", "/etc/pwx:/etc/pwx"},
-		[]string{ "/runc-entry-point.sh", "--debug" }, []string{"--upgrade-inplace"} )
+		[]string{"/runc-entry-point.sh", "--debug"}, []string{"--upgrade-inplace", "--info=stats2"}, reFilter)
 	if err != nil {
 		logrus.WithError(err).Error("Could not install ", imageName)
 		usage("Could not install " + imageName +
 			" - please inspect docker's log, and contact Portworx support.")
+	}
+
+	numTransferredFiles := -1
+	if len(reFilter.val) > 0 {
+		// get rid of ,'s: 25,455 -> 25455
+		matched := strings.Replace(string(reFilter.val), ",", "", -1)
+		numTransferredFiles, err = strconv.Atoi(matched)
+		if err != nil {
+			logrus.Warn("Could not find number of transferred files")
+			numTransferredFiles = -1 // reset, just in case
+		} else {
+			logrus.Infof("Number of transferred files: %d", numTransferredFiles)
+		}
 	}
 
 	logrus.Info("Installing Portworx...")
@@ -122,12 +191,27 @@ func installPxFromOciImage(di *utils.DockerInstaller, imageName string, cfg *uti
 
 	// TODO: Add Labels?
 
-	if err = runExternal(args[0], args[1:]...); err != nil {
+	var out cachingOutput
+	if err = runExternalWithOutput(&out, args[0], args[1:]...); err != nil {
 		logrus.WithError(err).Error("Could not install PX-RunC")
-		return err
+		return true, err
 	}
+	installOutput := out.String()
 
-	return nil
+	// figure out if update required
+	needsUpdate := false
+	if numTransferredFiles > 3 {
+		logrus.Infof("Portworx service restart required as %d files updated.", numTransferredFiles)
+		needsUpdate = true
+	} else if numTransferredFiles < 0 {
+		logrus.Info("Portworx service restart required as unknown number of files updated.")
+		needsUpdate = true
+	}
+	if isRestartRequired(installOutput) {
+		logrus.Info("Portworx service restart required due to configuration update.")
+		needsUpdate = true
+	}
+	return needsUpdate, nil
 }
 
 func validateMounts(mounts ...string) error {
@@ -142,7 +226,7 @@ func validateMounts(mounts ...string) error {
 		if err != nil {
 			return fmt.Errorf("File/Directory %s not found (%s) - please mount via 'run -v ...' option", m, err)
 		} else if st0.Dev == st1.Dev {
-			return fmt.Errorf("File/Directory %s not mounted - please mount via 'run -v ...' option", m, err)
+			return fmt.Errorf("File/Directory %s not mounted - please mount via 'run -v ...' option", m)
 		}
 	}
 	return nil
@@ -170,7 +254,7 @@ func doInstall() {
 	if err != nil {
 		logrus.WithError(err).Error("Could not talk to Docker")
 		usage("Could not talk to Docker" +
-				" - please restart using '-v /var/run/docker.sock:/var/run/docker.sock' option")
+			" - please restart using '-v /var/run/docker.sock:/var/run/docker.sock' option")
 	}
 
 	opts, err := di.ExtractConfig(id)
@@ -182,41 +266,57 @@ func doInstall() {
 
 	// TODO: Sanity checks for options
 	logrus.Debugf("OPTIONS:: %+v\n", opts)
-
-	err = installPxFromOciImage(di, pxImage, opts)
+	isRestartRequired, err := installPxFromOciImage(di, pxImage, opts)
 	if err != nil {
 		logrus.WithError(err).Error("Could not install Portworx service")
 		os.Exit(-1)
 	}
 
-	logrus.Info("Stopping Portworx service (if any)")
-	err = runExternal("/bin/sh", "-c", "systemctl stop portworx")
-	logrus.WithError(err).Debugf("Stopping done")
+	if isRestartRequired {
+		logrus.Warn("Restarting portworx service")
+		logrus.Info("Stopping Portworx service (if any)")
+		err = runExternal("/bin/sh", "-c", "systemctl stop portworx")
+		logrus.WithError(err).Debugf("Stopping done")
 
-	logrus.Info("Starting Portworx service")
-	err = runExternal("/bin/sh", "-c",
-		`systemctl daemon-reload && systemctl enable portworx && systemctl start portworx`)
-	if err != nil {
-		logrus.WithError(err).Error("Could not start Portworx service")
-		os.Exit(-1)
+		logrus.Info("Starting Portworx service")
+		err = runExternal("/bin/sh", "-c",
+			`systemctl daemon-reload && systemctl enable portworx && systemctl start portworx`)
+		if err != nil {
+			logrus.WithError(err).Error("Could not start Portworx service")
+			os.Exit(-1)
+		}
+	} else {
+		logrus.Info("Portworx service restart not required.")
 	}
 }
 
 func doUninstall() {
 	logrus.Info("Stopping Portworx service")
-	err := runExternal("/bin/sh", "-c",
+	var b bytes.Buffer
+	err := runExternalWithOutput(&b, "/bin/sh", "-c",
 		`systemctl stop portworx`)
 	if err != nil {
-		logrus.WithError(err).Error("Could not stop Portworx service")
-		os.Exit(-1) // NOTE: CRITICAL failure !!
+		strerr := b.String()
+		if strings.Contains(strerr, " not loaded") {
+			logrus.Info(strerr)
+		} else {
+			logrus.WithError(err).Error("Could not stop Portworx service")
+			os.Exit(-1) // NOTE: CRITICAL failure !!
+		}
 	}
 
 	logrus.Info("Disabling Portworx service")
-	err = runExternal("/bin/sh", "-c",
+	b.Reset()
+	err = runExternalWithOutput(&b, "/bin/sh", "-c",
 		`systemctl disable portworx`)
 	if err != nil {
-		logrus.WithError(err).Error("Could not disable Portworx service")
-		os.Exit(-1) // NOTE: CRITICAL failure !!
+		strerr := b.String()
+		if strings.Contains(strerr, "No such file or directory") {
+			logrus.Info("Portworx service already disabled")
+		} else {
+			logrus.WithError(err).Error("Could not disable Portworx service")
+			os.Exit(-1) // NOTE: CRITICAL failure !!
+		}
 	}
 
 	logrus.Info("Removing Portworx service bind-mount (if any)")
@@ -231,6 +331,82 @@ func doUninstall() {
 	if err != nil {
 		logrus.WithError(err).Warn("Could not remove all Portworx files")
 	}
+}
+
+// getKubernetesRootDir scans the external kubelet service for "--root-dir=XX" override, or returns a default kubelet dir
+func getKubernetesRootDir() (string, error) {
+	var out cachingOutput
+	args := strings.Fields(`/bin/ps --no-headers -o cmd -C kubelet`)
+	if err := runExternalWithOutput(&out, args[0], args[1:]...); err != nil {
+		err = fmt.Errorf("Could not find kubelet service: %s", err)
+		return "", err
+	}
+	m := xtractKubeletRegex.FindAllStringSubmatch(out.String(), -1)
+	if len(m) > 0 && len(m[0]) > 1 {
+		return m[0][1], nil
+	}
+	// return default value
+	return "/var/lib/kubelet", nil
+}
+
+// used for unit-tests
+var getKubernetesRootDirFn = getKubernetesRootDir
+
+// isRestartRequired returns FALSE if no updates detected, or if only POD-mounts -specific updates are present.
+// It returns TRUE if spec is new, if input not parseable, or non POD-mounts -specific updates present.
+func isRestartRequired(in string) bool {
+	if strings.Index(in, "SPEC UNCHANGED ") > 0 {
+		return false
+	} else if strings.Index(in, "SPEC CREATED ") > 0 {
+		return true
+	}
+
+	// find proper location for "/var/lib/kubelet/pods/"
+	kubeletPodsDir, err := getKubernetesRootDirFn()
+	if err != nil {
+		logrus.WithError(err).Error("Error scanning kubelet process")
+		return true
+	}
+	kubeletPodsDir += "/pods/"
+
+	search4 := " Updated mounts: add{"
+	addStartIdx := strings.Index(in, search4)
+	if addStartIdx > 0 {
+		addStartIdx += len(search4)
+		if endIdx := strings.Index(in[addStartIdx:], "}"); endIdx > 0 {
+			endIdx += addStartIdx
+			for _, p := range strings.Fields(in[addStartIdx:endIdx]) {
+				logrus.Debugf("add/%s/", p)
+				if !strings.HasPrefix(p, kubeletPodsDir) {
+					return true
+				}
+			}
+		} else {
+			logrus.Error("INTERNAL ERROR - found mounts add{ with no matching }")
+			return true
+		}
+	} else {
+		// reset -- we never found added mounts
+		addStartIdx = 0
+	}
+
+	search4 = " rm{"
+	if rmStartIdx := strings.Index(in[addStartIdx:], search4); rmStartIdx > 0 {
+		rmStartIdx += addStartIdx + len(search4)
+		if endIdx := strings.Index(in[rmStartIdx:], "}"); endIdx > 0 {
+			endIdx += rmStartIdx
+			for _, p := range strings.Fields(in[rmStartIdx:endIdx]) {
+				logrus.Debugf("rm/%s/", p)
+				if !strings.HasPrefix(p, kubeletPodsDir) {
+					return true
+				}
+			}
+		} else {
+			logrus.Error("INTERNAL ERROR - found mounts rm{ with no matching }")
+			return true
+		}
+	}
+	return false
 }
 
 func main() {
